requirements.txt
langchain
python-dotenv
langchain_community
pypdf
beautifulsoup4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
langchain-community
langchain-ollama
transformers
sentence-transformers
faiss-cpu
chromadb
langchain_chroma
streamlit